+++ src/screens/Avatars/AvatarScreen.tsx
@@ -1,6 +1,6 @@
 import { useGame } from '../../hooks/useGame.ts';
 import { Button, Container, Typography } from '@mui/material';
-import { ScreenType } from '../constants/screens.ts';
+import { ScreenType } from '../../constants/screens.ts';
 
 export const AvatarScreen = () => {
   const { setScreen } = useGame();
