+++ src/screens/Categories/CategoriesScreen.tsx
@@ -1,6 +1,6 @@
 import { useGame } from '../../hooks/useGame.ts';
 import { Button, Container, Typography } from '@mui/material';
-import { ScreenType } from '../constants/screens.ts';
+import { ScreenType } from '../../constants/screens.ts';
 
 export const CategoriesScreen = () => {
   const { setScreen } = useGame();
